Logic split + update readme
