+++ test/tezlib/listener.ts
@@ -2,7 +2,7 @@ import { TezosToolkit } from "@taquito/taquito";
 import { TonClient, abiContract } from "@tonclient/core";
 import { libNode } from "@tonclient/lib-node";
 import { TokenWallet } from "./everscale/everlib";
-import config from "./config.json";
+import config from "./listener.config.json";
 
 const RPC = "https://rpc.hangzhounet.teztnets.xyz";
 
@@ -16,23 +16,24 @@ const client = new TonClient({
 
 async function main() {
     const tezos = new TezosToolkit(RPC);
-    const everscale = new TokenWallet(client, null, config.everscale_token_listener);
+    const everscale = new TokenWallet(client, null, config.everscale.address);
     
-    console.log(`[!] EVERSCALE: Listening ${everscale.address} for incoming tokens...`)
-    console.log(`[!] TEZOS: Listening ${config.tezos_token_contract} for incoming tokens...`)
+    console.log(`[*] EVERSCALE: Listening ${everscale.address} for incoming tokens...`)
+    console.log(`[*] TEZOS: Listening ${config.tezos.token_contract} for incoming tokens...`)
+    console.log("\n");
 
     everscale.onTokenRecieved(data => {
-        console.log(`[!] EVERSCALE: На адрес ${everscale.address} пришло ${data.amount} токенов! Payload: ${data.payload}`);
+        console.log(`[*] EVERSCALE: Token recieved!\n | Amount: ${data.amount}\n | Payload: ${data.payload}`);
     })
 
     const tezosSubscription = tezos.stream.subscribeOperation({
-        and: [{destination: config.tezos_token_contract }]
+        and: [{destination: config.tezos.token_contract }]
     });
 
     tezosSubscription.on("data", (data: any) => {
         const [{string: reciever}, {int: amount}] = data.parameters.value.args[1].args;
-        if(reciever === config.tezos_token_listener) {
-            console.log(`[!] TEZOS: На адрес ${reciever} пришло ${amount} токенов!`);
+        if(reciever === config.tezos.address) {
+            console.log(`[*] TEZOS: Token recieved!\n | Amount: ${amount}`);
         }
     })
 }
