+++ test/tezlib/tezos_transfer.config.json
@@ -1,0 +1,29 @@
+{
+    "account": {
+        "pkh": "tz1gVYfPffnmhyZkiEXadeg5SS8uerbXo2DM",
+        "mnemonic": [
+            "myself",
+            "nothing",
+            "fury",
+            "disagree",
+            "uncover",
+            "embrace",
+            "custom",
+            "bike",
+            "frown",
+            "emotion",
+            "uncle",
+            "tissue",
+            "trip",
+            "render",
+            "lake"
+        ],
+        "email": "neaqjwkm.gcmmrqga@teztnets.xyz",
+        "password": "zTUenCLIRQ",
+        "amount": "81902732788",
+        "activation_code": "e859bdd3ce75b99abec30d1305cf84f520595f43"
+    },
+    "token_contract": "KT1MPrXmeAhHMdz9L3V25D38PndpT4cDwAi8",
+    "to": "tz1gVYfPffnmhyZkiEXadeg5SS8uerbXo2DM",
+    "amount": 1
+}
