+++ test/tezlib/transfer_everscale_token.ts
@@ -1,7 +1,7 @@
 import { TonClient } from "@tonclient/core";
 import { libNode } from "@tonclient/lib-node";
 import { TokenWallet } from "./everscale/everlib";
-import config from "./config.json";
+import config from "./everscale_transfer.config.json";
 
 TonClient.useBinaryLibrary(libNode);
 
@@ -13,17 +13,17 @@ const client = new TonClient({
 
 async function main() {
     const sender = new TokenWallet(client, {
-        public: config.everscale_token_sender.keys.public,
-        secret: config.everscale_token_sender.keys.secret
-    }, config.everscale_token_sender.address)
+        public: config.account.keys.public,
+        secret: config.account.keys.secret
+    }, config.account.address)
 
-    const reciever = new TokenWallet(client, null, config.everscale_token_listener)
+    const reciever = new TokenWallet(client, null, config.to)
 
     console.log("[!] Sender balance:", await sender.getBalance());
     console.log("[!] Reciever balance: ", await reciever.getBalance());
 
-    console.log(`[!] Sending ${config.everscale_token_sender.amount} tokens to ${reciever.address}`)
-    const transactionID = await sender.transfer(reciever.address, config.everscale_token_sender.amount, "address123");
+    console.log(`[!] Sending ${config.amount} tokens to ${reciever.address} with payload: ${config.transfer_payload}`)
+    const transactionID = await sender.transfer(reciever.address, config.amount, config.transfer_payload);
     console.log("[!] TXID:", transactionID)
 
     console.log("[!] Sender balance:", await sender.getBalance());
