+++ test/tezlib/transfer_tezos_token.ts
@@ -1,12 +1,12 @@
 import { Tezos } from "./tezos";
-import config from "./config.json";
+import config from "./tezos_transfer.config.json";
 
 const RPC = "https://rpc.hangzhounet.teztnets.xyz";
 
 async function main() {
-    const tezos = new Tezos(RPC, config.tezos_token_sender.keys)
-    console.log(`[!] Sending ${config.tezos_token_sender.amount} tokens to ${config.tezos_token_listener}`)
-    await tezos.transferToken(config.tezos_token_contract, config.tezos_token_listener, config.tezos_token_sender.amount)
+    const tezos = new Tezos(RPC, config.account)
+    console.log(`[!] Sending ${config.amount} tokens to ${config.to}`)
+    await tezos.transferToken(config.token_contract, config.to, config.amount)
     console.log("[!] Token sent succesfully!");
 }
